Refactored some of the testing code to their own components
