+++ src/components/Admin/StackedBarChart.tsx
@@ -1,0 +1,48 @@
+import React from "react"
+import {
+    BarChart,
+    Bar,
+    XAxis,
+    YAxis,
+    CartesianGrid,
+    Tooltip,
+    Legend
+  } from "recharts";
+  
+export type data = {
+    name:string,
+    approved:number,
+    denied:number
+  }
+
+type Props = {
+    listOfAdmins:data[]
+}
+
+const StackedBarChart:React.FC<Props> = (props) =>{
+
+    return(
+    <BarChart
+      width={500}
+      height={300}
+      data={props.listOfAdmins}
+      margin={{
+        top: 20,
+        right: 30,
+        left: 20,
+        bottom: 5
+      }}
+    >
+      <CartesianGrid strokeDasharray="3 3" />
+      <XAxis dataKey="name" />
+      <YAxis />
+      <Tooltip />
+      <Legend />
+      <Bar dataKey="approved" stackId="a" fill="#8884d8" />
+      <Bar dataKey="denied" stackId="a" fill="#82ca9d" />
+      
+    </BarChart>
+    )
+}
+
+export default StackedBarChart;
